+++ crates/tx3-cardano/src/coercion.rs
@@ -157,6 +157,16 @@ pub fn expr_into_address(
         ir::Expression::Hash(x) => policy_into_address(x, network),
         ir::Expression::Bytes(x) => bytes_into_address(x),
         ir::Expression::String(x) => string_into_address(x),
+        ir::Expression::EvalCompiler(x) => match x.as_ref() {
+            ir::CompilerOp::BuildScriptAddress(x) => {
+                let hash: primitives::Hash<28> = expr_into_hash(x)?;
+                policy_into_address(&hash.to_vec(), network)
+            }
+            _ => Err(Error::CoerceError(
+                format!("{:?}", expr),
+                "Address".to_string(),
+            )),
+        },
         _ => Err(Error::CoerceError(
             format!("{:?}", expr),
             "Address".to_string(),
@@ -177,6 +187,7 @@ pub fn expr_into_hash<const SIZE: usize>(
 ) -> Result<primitives::Hash<SIZE>, Error> {
     match ir {
         ir::Expression::Bytes(x) => Ok(primitives::Hash::from(x.as_slice())),
+        ir::Expression::Hash(x) => Ok(primitives::Hash::from(x.as_slice())),
         _ => Err(Error::CoerceError(format!("{:?}", ir), "Hash".to_string())),
     }
 }