+++ crates/tx3-lang/src/applying.rs
@@ -695,9 +695,6 @@ impl ir::AssetExpr {
 impl Composite for ir::Input {
     fn components(&self) -> Vec<&ir::Expression> {
         vec![&self.utxos, &self.redeemer]
-            .into_iter()
-            .chain(self.policy.iter().flat_map(|x| x.components()))
-            .collect()
     }
 
     fn try_map_components<F>(self, f: F) -> Result<Self, Error>
@@ -708,7 +705,6 @@ impl Composite for ir::Input {
             name: self.name,
             utxos: f(self.utxos)?,
             redeemer: f(self.redeemer)?,
-            policy: self.policy.map(|x| x.try_map_components(f)).transpose()?,
         })
     }
 }
@@ -838,10 +834,21 @@ impl Apply for ir::Expression {
             Self::EvalParam(x) => Ok(Self::EvalParam(Box::new(x.apply_args(args)?))),
             Self::EvalBuiltIn(x) => Ok(Self::EvalBuiltIn(Box::new(x.apply_args(args)?))),
             Self::EvalCoerce(x) => Ok(Self::EvalCoerce(Box::new(x.apply_args(args)?))),
+            Self::EvalCompiler(x) => Ok(Self::EvalCompiler(Box::new(x.apply_args(args)?))),
             Self::AdHocDirective(x) => Ok(Self::AdHocDirective(Box::new(x.apply_args(args)?))),
 
-            // it's safe to skip the remaining expressions as they are constant
-            _ => Ok(self),
+            // Don't fall into the temptation of simplifying the following cases under a single
+            // wildcard with a default implementation, it makes it really hard to detect missing
+            // implementation when adding new `Expression` variants.
+            Self::None => Ok(self),
+            Self::Bytes(_) => Ok(self),
+            Self::Number(_) => Ok(self),
+            Self::Bool(_) => Ok(self),
+            Self::String(_) => Ok(self),
+            Self::Address(_) => Ok(self),
+            Self::Hash(_) => Ok(self),
+            Self::UtxoRefs(_) => Ok(self),
+            Self::UtxoSet(_) => Ok(self),
         }
     }
 
@@ -865,10 +872,21 @@ impl Apply for ir::Expression {
             Self::EvalParam(x) => Ok(Self::EvalParam(Box::new(x.apply_inputs(args)?))),
             Self::EvalBuiltIn(x) => Ok(Self::EvalBuiltIn(Box::new(x.apply_inputs(args)?))),
             Self::EvalCoerce(x) => Ok(Self::EvalCoerce(Box::new(x.apply_inputs(args)?))),
+            Self::EvalCompiler(x) => Ok(Self::EvalCompiler(Box::new(x.apply_inputs(args)?))),
             Self::AdHocDirective(x) => Ok(Self::AdHocDirective(Box::new(x.apply_inputs(args)?))),
 
-            // it's safe to skip the remaining expressions as they are constant
-            _ => Ok(self),
+            // Don't fall into the temptation of simplifying the following cases under a single
+            // wildcard with a default implementation, it makes it really hard to detect missing
+            // implementation when adding new `Expression` variants.
+            Self::None => Ok(self),
+            Self::Bytes(_) => Ok(self),
+            Self::Number(_) => Ok(self),
+            Self::Bool(_) => Ok(self),
+            Self::String(_) => Ok(self),
+            Self::Address(_) => Ok(self),
+            Self::Hash(_) => Ok(self),
+            Self::UtxoRefs(_) => Ok(self),
+            Self::UtxoSet(_) => Ok(self),
         }
     }
 
@@ -892,10 +910,21 @@ impl Apply for ir::Expression {
             Self::EvalParam(x) => Ok(Self::EvalParam(Box::new(x.apply_fees(fees)?))),
             Self::EvalBuiltIn(x) => Ok(Self::EvalBuiltIn(Box::new(x.apply_fees(fees)?))),
             Self::EvalCoerce(x) => Ok(Self::EvalCoerce(Box::new(x.apply_fees(fees)?))),
+            Self::EvalCompiler(x) => Ok(Self::EvalCompiler(Box::new(x.apply_fees(fees)?))),
             Self::AdHocDirective(x) => Ok(Self::AdHocDirective(Box::new(x.apply_fees(fees)?))),
 
-            // it's safe to skip the remaining expressions as they are constant
-            _ => Ok(self),
+            // Don't fall into the temptation of simplifying the following cases under a single
+            // wildcard with a default implementation, it makes it really hard to detect missing
+            // implementation when adding new `Expression` variants.
+            Self::None => Ok(self),
+            Self::Bytes(_) => Ok(self),
+            Self::Number(_) => Ok(self),
+            Self::Bool(_) => Ok(self),
+            Self::String(_) => Ok(self),
+            Self::Address(_) => Ok(self),
+            Self::Hash(_) => Ok(self),
+            Self::UtxoRefs(_) => Ok(self),
+            Self::UtxoSet(_) => Ok(self),
         }
     }
 
@@ -908,10 +937,21 @@ impl Apply for ir::Expression {
             Self::EvalParam(x) => x.is_constant(),
             Self::EvalBuiltIn(x) => x.is_constant(),
             Self::EvalCoerce(x) => x.is_constant(),
+            Self::EvalCompiler(x) => x.is_constant(),
             Self::AdHocDirective(x) => x.is_constant(),
 
-            // it's safe to skip the remaining expressions as they are constant
-            _ => true,
+            // Don't fall into the temptation of simplifying the following cases under a single
+            // wildcard with a default implementation, it makes it really hard to detect missing
+            // implementation when adding new `Expression` variants.
+            Self::None => true,
+            Self::Bytes(_) => true,
+            Self::Number(_) => true,
+            Self::Bool(_) => true,
+            Self::String(_) => true,
+            Self::Address(_) => true,
+            Self::Hash(_) => true,
+            Self::UtxoRefs(_) => true,
+            Self::UtxoSet(_) => true,
         }
     }
 
@@ -924,10 +964,21 @@ impl Apply for ir::Expression {
             Self::EvalParam(x) => x.params(),
             Self::EvalBuiltIn(x) => x.params(),
             Self::EvalCoerce(x) => x.params(),
+            Self::EvalCompiler(x) => x.params(),
             Self::AdHocDirective(x) => x.params(),
 
-            // it's safe to skip the remaining expressions as they are constant
-            _ => BTreeMap::new(),
+            // Don't fall into the temptation of simplifying the following cases under a single
+            // wildcard with a default implementation, it makes it really hard to detect missing
+            // implementation when adding new `Expression` variants.
+            Self::None => BTreeMap::new(),
+            Self::Bytes(_) => BTreeMap::new(),
+            Self::Number(_) => BTreeMap::new(),
+            Self::Bool(_) => BTreeMap::new(),
+            Self::String(_) => BTreeMap::new(),
+            Self::Address(_) => BTreeMap::new(),
+            Self::Hash(_) => BTreeMap::new(),
+            Self::UtxoRefs(_) => BTreeMap::new(),
+            Self::UtxoSet(_) => BTreeMap::new(),
         }
     }
 
@@ -943,10 +994,21 @@ impl Apply for ir::Expression {
             Self::EvalParam(x) => x.queries(),
             Self::EvalBuiltIn(x) => x.queries(),
             Self::EvalCoerce(x) => x.queries(),
+            Self::EvalCompiler(x) => x.queries(),
             Self::AdHocDirective(x) => x.queries(),
 
-            // it's safe to skip the remaining expressions as they are constant
-            _ => BTreeMap::new(),
+            // Don't fall into the temptation of simplifying the following cases under a single
+            // wildcard with a default implementation, it makes it really hard to detect missing
+            // implementation when adding new `Expression` variants.
+            Self::None => BTreeMap::new(),
+            Self::Bytes(_) => BTreeMap::new(),
+            Self::Number(_) => BTreeMap::new(),
+            Self::Bool(_) => BTreeMap::new(),
+            Self::String(_) => BTreeMap::new(),
+            Self::Address(_) => BTreeMap::new(),
+            Self::Hash(_) => BTreeMap::new(),
+            Self::UtxoRefs(_) => BTreeMap::new(),
+            Self::UtxoSet(_) => BTreeMap::new(),
         }
     }
 
@@ -965,6 +1027,7 @@ impl Apply for ir::Expression {
                     .map(|x| x.reduce())
                     .collect::<Result<_, _>>()?,
             )),
+            ir::Expression::EvalCompiler(x) => Ok(Self::EvalCompiler(Box::new(x.reduce()?))),
             ir::Expression::AdHocDirective(x) => Ok(Self::AdHocDirective(Box::new(x.reduce()?))),
 
             // the following ones can be turned into simpler expressions
@@ -980,7 +1043,19 @@ impl Apply for ir::Expression {
                 ir::Param::Set(x) => Ok(x),
                 x => Ok(ir::Expression::EvalParam(Box::new(x.reduce()?))),
             },
-            _ => Ok(self),
+
+            // Don't fall into the temptation of simplifying the following cases under a single
+            // wildcard with a default implementation, it makes it really hard to detect missing
+            // implementation when adding new `Expression` variants.
+            Self::None => Ok(self),
+            Self::Bytes(_) => Ok(self),
+            Self::Number(_) => Ok(self),
+            Self::Bool(_) => Ok(self),
+            Self::String(_) => Ok(self),
+            Self::Address(_) => Ok(self),
+            Self::Hash(_) => Ok(self),
+            Self::UtxoRefs(_) => Ok(self),
+            Self::UtxoSet(_) => Ok(self),
         }
     }
 }
@@ -1038,6 +1113,23 @@ impl Composite for ir::AdHocDirective {
     }
 }
 
+impl Composite for ir::CompilerOp {
+    fn components(&self) -> Vec<&ir::Expression> {
+        match self {
+            ir::CompilerOp::BuildScriptAddress(x) => vec![x],
+        }
+    }
+
+    fn try_map_components<F>(self, f: F) -> Result<Self, Error>
+    where
+        F: Fn(ir::Expression) -> Result<ir::Expression, Error> + Clone,
+    {
+        match self {
+            ir::CompilerOp::BuildScriptAddress(x) => Ok(ir::CompilerOp::BuildScriptAddress(f(x)?)),
+        }
+    }
+}
+
 impl Composite for ir::Signers {
     fn components(&self) -> Vec<&ir::Expression> {
         self.signers.iter().collect()
@@ -1055,7 +1147,7 @@ impl Composite for ir::Signers {
 
 impl Composite for ir::Collateral {
     fn components(&self) -> Vec<&ir::Expression> {
-        self.query.components()
+        vec![&self.utxos]
     }
 
     fn try_map_components<F>(self, f: F) -> Result<Self, Error>
@@ -1063,7 +1155,7 @@ impl Composite for ir::Collateral {
         F: Fn(ir::Expression) -> Result<ir::Expression, Error> + Clone,
     {
         Ok(Self {
-            query: self.query.try_map_components(f)?,
+            utxos: f(self.utxos)?,
         })
     }
 }