+++ crates/tx3-lang/src/ast.rs
@@ -263,7 +263,7 @@ impl HexStringLiteral {
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub enum CollateralBlockField {
-    From(AddressExpr),
+    From(DataExpr),
     MinAmount(DataExpr),
     Ref(DataExpr),
 }
@@ -274,13 +274,6 @@ impl CollateralBlockField {
             CollateralBlockField::From(_) => "from",
             CollateralBlockField::MinAmount(_) => "min_amount",
             CollateralBlockField::Ref(_) => "ref",
-        }
-    }
-
-    pub fn as_address_expr(&self) -> Option<&AddressExpr> {
-        match self {
-            CollateralBlockField::From(x) => Some(x),
-            _ => None,
         }
     }
 
@@ -306,7 +299,7 @@ impl CollateralBlock {
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub enum InputBlockField {
-    From(AddressExpr),
+    From(DataExpr),
     DatumIs(Type),
     MinAmount(DataExpr),
     Redeemer(DataExpr),
@@ -321,13 +314,6 @@ impl InputBlockField {
             InputBlockField::MinAmount(_) => "min_amount",
             InputBlockField::Redeemer(_) => "redeemer",
             InputBlockField::Ref(_) => "ref",
-        }
-    }
-
-    pub fn as_address_expr(&self) -> Option<&AddressExpr> {
-        match self {
-            InputBlockField::From(x) => Some(x),
-            _ => None,
         }
     }
 
@@ -387,7 +373,7 @@ impl InputBlock {
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub enum OutputBlockField {
-    To(Box<AddressExpr>),
+    To(Box<DataExpr>),
     Amount(Box<DataExpr>),
     Datum(Box<DataExpr>),
 }