+++ crates/tx3-lang/src/ir.rs
@@ -15,7 +15,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::{Utxo, UtxoRef};
 
-pub const IR_VERSION: &str = "v1alpha6";
+pub const IR_VERSION: &str = "v1alpha7";
 
 #[derive(Encode, Decode, Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
 pub struct StructExpr {
@@ -42,6 +42,14 @@ pub enum Coerce {
 
 pub type PropertyIndex = usize;
 
+/// Operations that are executed during the "apply" phase.
+///
+/// These are operations that are executed during the "apply" phase, as opposed
+/// to the compiler operations that are executed during the "compile" phase.
+///
+/// These ops can be executed (aka "reduced") very early in the process. As long
+/// as they underlying expressions are "constant" (aka: don't rely on external
+/// data), the will be simplified directly during the "apply" phase.
 #[derive(Encode, Decode, Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
 pub enum BuiltInOp {
     NoOp(Expression),
@@ -49,6 +57,19 @@ pub enum BuiltInOp {
     Sub(Expression, Expression),
     Negate(Expression),
     Property(Expression, PropertyIndex),
+}
+
+/// Operations that are performed by the compiler.
+///
+/// These are operations that are performed by the compiler, as opposed to the
+/// built-in operations that are executed (aka "reduced") during the "apply"
+/// phase.
+///
+/// These ops can't be executed earlier because they are either: chain-specific
+/// or rely on data that is only available to the compiler.
+#[derive(Encode, Decode, Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
+pub enum CompilerOp {
+    BuildScriptAddress(Expression),
 }
 
 #[derive(Encode, Decode, Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
@@ -179,6 +200,7 @@ pub enum Expression {
     MinUtxo(Box<Expression>),
     EvalParam(Box<Param>),
     EvalBuiltIn(Box<BuiltInOp>),
+    EvalCompiler(Box<CompilerOp>),
     EvalCoerce(Box<Coerce>),
 
     // pass-through
@@ -210,6 +232,12 @@ impl From<BuiltInOp> for Expression {
     }
 }
 
+impl From<CompilerOp> for Expression {
+    fn from(op: CompilerOp) -> Self {
+        Self::EvalCompiler(Box::new(op))
+    }
+}
+
 impl From<Coerce> for Expression {
     fn from(coerce: Coerce) -> Self {
         Self::EvalCoerce(Box::new(coerce))
@@ -234,10 +262,6 @@ pub struct Input {
     pub name: String,
     pub utxos: Expression,
     pub redeemer: Expression,
-
-    // TODO: revisit this value. This is meant to support adding reference scripts by pointing to a
-    // policy that specifies the ref. This is not straightforward and hard to reason about.
-    pub policy: Option<PolicyExpr>,
 }
 
 #[derive(Encode, Decode, Serialize, Deserialize, Debug, Clone)]
@@ -261,7 +285,7 @@ pub struct Mint {
 
 #[derive(Encode, Decode, Serialize, Deserialize, Debug, Clone)]
 pub struct Collateral {
-    pub query: InputQuery,
+    pub utxos: Expression,
 }
 
 #[derive(Encode, Decode, Serialize, Deserialize, Debug, Clone)]